refactor: restructure the command line interface
Use a single `Command` type to handle all of the commands.

Move the version type to its own `semver` package and don't use a global
variable to store the version information.

Move the main invokation of the program to the main file as Go documentation
suggests. The project has only one executable so there is no need for the `cmd`
package.